requirements.txt
dash==2.17.1
dash_bootstrap_components==1.6.0
dash_core_components==2.0.0
dash_html_components==2.0.0
dash_leaflet==1.0.15
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
gunicorn
dash-tools
